Refactor rankings logic into separate module and add Vercel serverless support
